Django_FCM_package/pyproject.toml
[build-system]
requires = ["setuptools>=58"]
build-backend = "setuptools.build_meta"

[project]
name = "fcm-messaging"
version = "1.0.0"
description = "FCM Messaging Package for Django"
requires-python = ">=3.8.20"
dependencies = [
    "Django>=4.2",
    "djangorestframework>=3.15.0",
    "psycopg2-binary==2.9.10",
    "firebase-admin>=5.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=6.2.5"]

[tool.setuptools]
packages = ['FCM_package', 'api']

# [tool.setuptools.plugins]
# "pyproject.toml" = "setuptools.build_meta"
